email-to-kg/requirements.txt
anthropic>=0.18.0
pyyaml>=6.0
python-dateutil>=2.8.2
